# docs/api/openapi.yml
openapi: 3.0.0
info:
    title: WildDuck API
    description: WildDuck API docs
    version: 1.0.0
    contact:
        url: 'https://github.com/nodemailer/wildduck'
servers:
    - url: 'https://api.wildduck.email'
...
paths:
...
    '/webhooks/{webhook}':
        delete:
            tags:
                - Webhooks
            summary: Delete a webhook
            operationId: deleteWebhook
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SuccessResponse'
        parameters:
            - name: webhook
              in: path
              description: ID of the Webhook
              required: true
              schema:
                  type: string
    /webhooks:
        get:
            tags:
                - Webhooks
            summary: List registered Webhooks
            operationId: getWebhooks
            parameters:
                - name: type
                  in: query
                  description: 'Prefix or exact match. Prefix match must end with ".*", eg "channel.*". Use "*" for all types'
                  required: true
                  schema:
                      type: string
                - name: user
                  in: query
                  description: User ID
                  schema:
                      type: string
                - name: limit
                  in: query
                  description: How many records to return
                  schema:
                      type: number
                - name: page
                  in: query
                  description: 'Current page number. Informational only, page numbers start from 1'
                  schema:
                      type: number
                - name: next
                  in: query
                  description: 'Cursor value for next page, retrieved from nextCursor response value'
                  schema:
                      type: number
                - name: previous
                  in: query
                  description: 'Cursor value for previous page, retrieved from previousCursor response value'
                  schema:
                      type: number
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetWebhooksResponse'
        post:
            tags:
                - Webhooks
            summary: Create new Webhook
            description: Create new webhook
            operationId: createWebhook
            requestBody:
                description: Create new webhook
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateWebhookRequest'
                required: true
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateWebhookResponse'
components:
    schemas:
...
        SuccessResponse:
            required:
                - success
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
...
        GetWebhooksResponse:
            required:
                - success
                - total
                - page
                - previousCursor
                - nextCursor
                - results
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                total:
                    type: number
                    description: How many results were found
                page:
                    type: number
                    description: Current page number. Derived from page query argument
                previousCursor:
                    type: string
                    description: Either a cursor string or false if there are not any previous results
                nextCursor:
                    type: string
                    description: Either a cursor string or false if there are not any next results
                results:
                    type: array
                    items:
                        $ref: '#/components/schemas/GetWebhooksResult'
                    description: Webhook listing
        CreateWebhookResponse:
            required:
                - success
                - id
            type: object
            properties:
                success:
                    type: boolean
                    description: Indicates successful response
                id:
                    type: string
                    description: ID of the Address
...
        CreateWebhookRequest:
            required:
                - type
                - url
            type: object
            properties:
                type:
                    type: array
                    items:
                        type: string
                    description: An array of event types to match. For prefix match use ".*" at the end (eg. "user.*") or "*" for all types
                user:
                    type: string
                    description: User ID to match (only makes sense for user specific resources)
                url:
                    type: string
                    description: URL to POST data to
...
        GetWebhooksResult:
            required:
                - id
                - type
                - user
                - url
            type: object
            properties:
                id:
                    type: string
                    description: Webhooks unique ID (24 byte hex)
                type:
                    type: array
                    items:
                        type: string
                    description: An array of event types this webhook matches
                user:
                    type: string
                    description: User ID or null
                url:
                    type: string
                    description: Webhook URL
    securitySchemes:
        AccessTokenAuth:
            name: X-Access-Token
            type: apiKey
            in: header
            description: |-
                If authentication is enabled in the WildDuck configuration, you will need to supply an access token in the `X-Access-Token` header.

                ```json
                {
                  "X-Access-Token": "59fc66a03e54454869460e45"
                }
                ```
security:
    - AccessTokenAuth: []